msm: clock-8960: Switch adm to ebi1_clk
Before the bus driver was voting on ebi1_clk we needed to have
adm use a dummy clock so that ebi1 wasn't voted off when adm no
longer used ebi1. Now that the bus driver is voting on ebi1
correctly, remove this workaround so that ebi1 stays out of self
refresh when a data mover transaction is occurring during power
collapse.

Change-Id: I74e6209271aa2a18e40c972ead8325c70c11e836
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>