ARM: 7293/1: logical_cpu_map: decouple CPU mapping from SMP
It turns out that the logical CPU mapping is useful even when !CONFIG_SMP
for manipulation of devices like interrupt and power controllers when
running a UP kernel on a CPU other than 0. This can happen when kexecing
a UP image from an SMP kernel.

In the future, multi-cluster systems running AMP configurations will
require something similar for mapping cluster IDs, so it makes sense to
decouple this logic in preparation for this support.

Acked-by: Yang Bai <hamo.by@gmail.com>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Reported-by: Joerg Roedel <joerg.roedel@amd.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>

Conflicts:
	arch/arm/mach-exynos/hotplug.c
	arch/arm/mach-exynos/platsmp.c
	arch/arm/mach-highbank/highbank.c
	arch/arm/mach-imx/src.c
	arch/arm/mach-msm/hotplug.c
	arch/arm/mach-realview/hotplug.c
	arch/arm/mach-shmobile/smp-r8a7779.c
	arch/arm/mach-shmobile/smp-sh73a0.c
	arch/arm/mach-ux500/hotplug.c
	arch/arm/mach-ux500/platsmp.c
	arch/arm/mach-vexpress/hotplug.c
	arch/arm/plat-versatile/platsmp.c