msm: smd: Prevent sending interrupt if port is closed
If a client is in the middle of a write and the SMD
port is closed, reset the data-written flag to prevent
sending an interrupt to the remote processor.  This interrupt
is harmless and should be ignored by the remote processor
but in some architectures, this interrupt can cause
problems with power collapse resulting in a bus lock.

CRs-Fixed: 334822
Change-Id: Iab738b06af53a538cbf2956807baea2839e05355
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>